Move TableColumnDelegatesClickAndKeydown
